// src/shared/api/generated/__swagger_client/models/EmployeeDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeDto = {
    id?: string;
    personId?: string;
    lastName?: string;
    firstName?: string;
    middleName?: string;
    organizationId?: string;
    organizationName?: string;
    orgUnitId?: string;
    orgUnitName?: string;
    personnelNumber?: string;
    positionName?: string;
    rankName?: string;
    hireDate?: string;
    dismissalDate?: string;
    responsible?: boolean;
    active?: boolean;
};